+++ examples/typescript/http-server-pool/express-hybrid/src/types.ts
@@ -1,6 +1,7 @@
-import type { ThreadPoolOptions } from 'poolifier'
+import type { Worker } from 'worker_threads'
+import type { PoolOptions } from 'poolifier'
 
-export interface ClusterWorkerData extends ThreadPoolOptions {
+export interface ClusterWorkerData extends PoolOptions<Worker> {
   port: number
   workerFile: string
   minWorkers?: number