+++ examples/typescript/http-server-pool/fastify-hybrid/src/types.ts
@@ -1,4 +1,5 @@
-import type { ThreadPoolOptions } from 'poolifier'
+import type { Worker } from 'worker_threads'
+import type { PoolOptions } from 'poolifier'
 
 export interface ClusterWorkerData extends FastifyPoolifierOptions {
   port: number
@@ -21,7 +22,7 @@ export interface ThreadWorkerResponse<T = unknown> {
   data: T
 }
 
-export interface FastifyPoolifierOptions extends ThreadPoolOptions {
+export interface FastifyPoolifierOptions extends PoolOptions<Worker> {
   workerFile: string
   minWorkers?: number
   maxWorkers?: number