+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.1] - 2023-12-16
+
+### Fixed
+
+- Fix pool options TS type definition.
+
+## [3.1.0] - 2023-12-16
+
+### Changed
+
+- TypeScript breaking change: merge ThreadPoolOptions and ClusterPoolOptions types into PoolOptions type.
+
 ## [3.0.14] - 2023-12-13
 
 ### Fixed